fixed displaying removing channel dialog

+++ app/src/main/kotlin/mikhail/shell/video/hosting/presentation/channel/screen/sections/ChannelHeader.kt
@@ -456,15 +456,17 @@ fun ChannelActionsButton(
                 actionDialogVisible = false
             }
         )
-        Dialog(
-            onSubmit = {
-                onRemove(channelId)
-            },
-            onDismiss = {
-                removeDialogVisible = false
-            },
-            dialogTitle = "Удаление канала",
-            dialogDescription = "Вы действительно хотите удалить канал?"
-        )
+        if (removeDialogVisible) {
+            Dialog(
+                onSubmit = {
+                    onRemove(channelId)
+                },
+                onDismiss = {
+                    removeDialogVisible = false
+                },
+                dialogTitle = "Удаление канала",
+                dialogDescription = "Вы действительно хотите удалить канал?"
+            )
+        }
     }
 }
